app/src/data_gen/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions~=1.24.0
azure-storage-blob~=12.28.0
azure-identity~=1.25.2
